+++ test/cases/stream/data/data.go
@@ -1,0 +1,131 @@
+package data
+
+import (
+	"context"
+	"embed"
+	"encoding/base64"
+	"encoding/json"
+	"io"
+	"strconv"
+	"time"
+
+	"github.com/google/go-cmp/cmp"
+	g "github.com/onsi/ginkgo/v2"
+	gm "github.com/onsi/gomega"
+	grpclib "google.golang.org/grpc"
+	"google.golang.org/protobuf/encoding/protojson"
+	"google.golang.org/protobuf/testing/protocmp"
+	"google.golang.org/protobuf/types/known/timestamppb"
+	"sigs.k8s.io/yaml"
+
+	common_v1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
+	model_v1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
+	stream_v1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/stream/v1"
+	"github.com/apache/skywalking-banyandb/pkg/test/helpers"
+)
+
+//go:embed input/*.yaml
+var inputFS embed.FS
+
+//go:embed want/*.yaml
+var wantFS embed.FS
+
+//go:embed testdata/*.json
+var dataFS embed.FS
+
+// VerifyFn verify whether the query response matches the wanted result
+var VerifyFn = func(sharedContext helpers.SharedContext, args helpers.Args) {
+	i, err := inputFS.ReadFile("input/" + args.Input + ".yaml")
+	gm.Expect(err).NotTo(gm.HaveOccurred())
+	query := &stream_v1.QueryRequest{}
+	helpers.UnmarshalYAML(i, query)
+	query.TimeRange = helpers.TimeRange(args, sharedContext)
+	c := stream_v1.NewStreamServiceClient(sharedContext.Connection)
+	ctx := context.Background()
+	resp, err := c.Query(ctx, query)
+	if args.WantErr {
+		if err == nil {
+			g.Fail("expect error")
+		}
+		return
+	}
+	gm.Expect(err).NotTo(gm.HaveOccurred(), query.String())
+	if args.WantEmpty {
+		gm.Expect(resp.Elements).To(gm.BeEmpty())
+		return
+	}
+	if args.Want == "" {
+		args.Want = args.Input
+	}
+	ww, err := wantFS.ReadFile("want/" + args.Want + ".yaml")
+	gm.Expect(err).NotTo(gm.HaveOccurred())
+	want := &stream_v1.QueryResponse{}
+	helpers.UnmarshalYAML(ww, want)
+	gm.Expect(cmp.Equal(resp, want,
+		protocmp.IgnoreUnknown(),
+		protocmp.IgnoreFields(&stream_v1.Element{}, "timestamp"),
+		protocmp.Transform())).
+		To(gm.BeTrue(), func() string {
+			j, err := protojson.Marshal(resp)
+			if err != nil {
+				return err.Error()
+			}
+			y, err := yaml.JSONToYAML(j)
+			if err != nil {
+				return err.Error()
+			}
+			return string(y)
+		})
+}
+
+func loadData(stream stream_v1.StreamService_WriteClient, dataFile string, baseTime time.Time, interval time.Duration) {
+	var templates []interface{}
+	content, err := dataFS.ReadFile("testdata/" + dataFile)
+	gm.Expect(err).ShouldNot(gm.HaveOccurred())
+	gm.Expect(json.Unmarshal(content, &templates)).ShouldNot(gm.HaveOccurred())
+	bb, _ := base64.StdEncoding.DecodeString("YWJjMTIzIT8kKiYoKSctPUB+")
+	for i, template := range templates {
+		rawSearchTagFamily, errMarshal := json.Marshal(template)
+		gm.Expect(errMarshal).ShouldNot(gm.HaveOccurred())
+		searchTagFamily := &model_v1.TagFamilyForWrite{}
+		gm.Expect(protojson.Unmarshal(rawSearchTagFamily, searchTagFamily)).ShouldNot(gm.HaveOccurred())
+		e := &stream_v1.ElementValue{
+			ElementId: strconv.Itoa(i),
+			Timestamp: timestamppb.New(baseTime.Add(interval * time.Duration(i))),
+			TagFamilies: []*model_v1.TagFamilyForWrite{
+				{
+					Tags: []*model_v1.TagValue{
+						{
+							Value: &model_v1.TagValue_BinaryData{
+								BinaryData: bb,
+							},
+						},
+					},
+				},
+			},
+		}
+		e.TagFamilies = append(e.TagFamilies, searchTagFamily)
+		errInner := stream.Send(&stream_v1.WriteRequest{
+			Metadata: &common_v1.Metadata{
+				Name:  "sw",
+				Group: "default",
+			},
+			Element: e,
+		})
+		gm.Expect(errInner).ShouldNot(gm.HaveOccurred())
+	}
+}
+
+// Write data into the server
+func Write(conn *grpclib.ClientConn, dataFile string, baseTime time.Time, interval time.Duration) {
+	c := stream_v1.NewStreamServiceClient(conn)
+	ctx := context.Background()
+	writeClient, err := c.Write(ctx)
+	gm.Expect(err).NotTo(gm.HaveOccurred())
+	loadData(writeClient, dataFile, baseTime, interval)
+	gm.Expect(writeClient.CloseSend()).To(gm.Succeed())
+	gm.Eventually(func() error {
+		_, err := writeClient.Recv()
+		return err
+	}).Should(gm.Equal(io.EOF))
+}
