+++ test/cases/stream/data/input/all.yaml
