+++ test/cases/stream/data/input/filter_tag.yaml
