+++ test/cases/stream/data/input/filter_tag_empty.yaml
