+++ test/cases/stream/data/input/global_index.yaml
