+++ test/cases/stream/data/input/having.yaml
