+++ test/cases/stream/data/input/less.yaml
