+++ test/cases/stream/data/input/less_eq.yaml
