+++ test/cases/stream/data/input/logical.yaml
