+++ test/cases/stream/data/input/offset.yaml
