+++ test/cases/stream/data/input/search.yaml
