+++ test/cases/stream/data/want/all.yaml
