+++ test/cases/stream/data/want/filter_tag.yaml
