+++ test/cases/stream/data/want/global_index.yaml
