+++ test/cases/stream/data/want/having.yaml
