+++ test/cases/stream/data/want/less.yaml
