+++ test/cases/stream/data/want/less_eq.yaml
