+++ test/cases/stream/data/want/logical.yaml
