+++ test/cases/stream/data/want/offset.yaml
