+++ test/cases/stream/data/want/search.yaml
