+++ test/integration/query/query_suite_test.go
@@ -28,6 +28,7 @@ import (
 	cases_measure "github.com/apache/skywalking-banyandb/test/cases/measure"
 	cases_measure_data "github.com/apache/skywalking-banyandb/test/cases/measure/data"
 	cases_stream "github.com/apache/skywalking-banyandb/test/cases/stream"
+	cases_stream_data "github.com/apache/skywalking-banyandb/test/cases/stream/data"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	grpclib "google.golang.org/grpc"
@@ -51,7 +52,7 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 		Level: "warn",
 	})).To(Succeed())
 	var addr string
-	addr, deferFunc = setup.SetUp()
+	addr, _, deferFunc = setup.SetUp()
 	conn, err := grpclib.Dial(
 		addr,
 		grpclib.WithTransportCredentials(insecure.NewCredentials()),
@@ -59,7 +60,7 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 	Expect(err).NotTo(HaveOccurred())
 	now = timestamp.NowMilli()
 	interval := 500 * time.Millisecond
-	cases_stream.Write(conn, "data.json", now, interval)
+	cases_stream_data.Write(conn, "data.json", now, interval)
 	cases_measure_data.Write(conn, "service_traffic", "sw_metric", "service_traffic_data.json", now, interval)
 	cases_measure_data.Write(conn, "service_instance_traffic", "sw_metric", "service_instance_traffic_data.json", now, interval)
 	cases_measure_data.Write(conn, "service_cpm_minute", "sw_metric", "service_cpm_minute_data.json", now, interval)
